+++ server/middlewares/checkAuth.js
